实现行为树

+++ SDK/sdk_macros.h
@@ -10,7 +10,7 @@
 #define SDK_MAX(a, b) ((a)>(b)?(a):(b))
 
 #define SDK_CONTAINER_OF(ptr, type, field) \
-  ((type *) ((char *) (ptr) - ((unsigned long) &((type *) 0)->field)))
+  ((type *) ((char *) (ptr) - ((size_t) &((type *) 0)->field)))
 
 
 #define SDK_PAGES(T,U) (((T)-1)/(U) + 1)
